% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.r
\name{addCycle}
\alias{addCycle}
\title{Calculate the PAWMAP cycle based on sampling date}
\usage{
addCycle(dateField)
}
\arguments{
\item{dateField}{A vector if sampling dates}
}
\value{
A vector of cycle numbers for each sampling date
}
\description{
Calculate the PAWMAP cycle based on sampling date
}
