% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFishTraits.r
\name{addFishTraits}
\alias{addFishTraits}
\title{Add fish traits to a data frame of fish counts}
\usage{
addFishTraits(
  df,
  speciesIn = "common_name",
  sppLook = "Common_Name",
  traits = c("Family", "Species_Name", "Common_Name", "Origin"),
  filterNA = TRUE
)
}
\arguments{
\item{df}{The data frame containing fish counts}

\item{speciesIn}{The field in df that identifies species}

\item{sppLook}{Look up species traits by: "Common_Name" (default)
or "Species_Name"}

\item{traits}{Which trait fields should be included?}

\item{filterNA}{Should species not in the traits table be dropped? (A warning
listing the unmatched species is provided.)}
}
\value{
A table of fish species presence (number of surveys on which
species was captured), and abundance (total number of individuals captured)
}
\description{
Add fish traits to a data frame of fish counts
}
