% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodge_v.r
\name{collidev}
\alias{collidev}
\title{Allow vertical dodging in ggplot}
\usage{
collidev(data, height = NULL, name, strategy, check.height = TRUE)
}
\description{
Detect and prevent collisions.
Powers dodging, stacking and filling.
}
