% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Panel_Table.r
\name{create_Panel_Table}
\alias{create_Panel_Table}
\title{Create table of PAWMAP sampling years, panels and cycles}
\usage{
create_Panel_Table(cycles = 5)
}
\arguments{
\item{cycles}{The number of four year rotational panels to include in the table}
}
\value{
A table of PAWMAP sampling years, panels and cycles
}
\description{
Create table of PAWMAP sampling years, panels and cycles
}
