% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.r
\name{.simpleCap}
\alias{.simpleCap}
\title{A Set of Miscellaneous Functions for Working with Data.}
\usage{
.simpleCap(x)
}
\arguments{
\item{x}{A string to convert to title case.}
}
\value{
A string in title case
}
\description{
A Set of Miscellaneous Functions for Working with Data.
}
\examples{
.simpleCap('a lower case string')
}
