% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixFishNames.r
\name{dropAmphibs}
\alias{dropAmphibs}
\title{Drop Amphibians from a data frame of fish names}
\usage{
dropAmphibs(dfm, vbl)
}
\arguments{
\item{dfm}{The data frame containing the fish data}

\item{vbl}{A field within dfm containing the fish names}
}
\value{
A A data frame without amphibians
}
\description{
Drop Amphibians from a data frame of fish names
}
