% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.r
\name{fitStats}
\alias{fitStats}
\title{Create text summarizing regression fit}
\usage{
fitStats(fit)
}
\arguments{
\item{fit}{fit object of class lm}
}
\value{
text with adj. r^2, p, slope and intercept summaries
}
\description{
Create text summarizing regression fit
}

