% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixFishNames.r
\name{fixFishNames}
\alias{fixFishNames}
\title{Standardize fish names and correct common misspellings}
\usage{
fixFishNames(vct)
}
\arguments{
\item{vct}{A vector containing the fish names}
}
\value{
A A vector with corrected fish names
}
\description{
Standardize fish names and correct common misspellings
}

