% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmp1_2_metadata}
\alias{hmp1_2_metadata}
\title{Public metadata of HMP1-II study
Provides sample information and body sites}
\format{matrix}
\usage{
hmp1_2_metadata
}
\description{
Public metadata of HMP1-II study
Provides sample information and body sites
}
\references{
\url{http://github.com/philippmuench}
}
\keyword{datasets}
