% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmp1_2_metaphlan}
\alias{hmp1_2_metaphlan}
\title{Metaphlan2 HMP1-II
Can be loaded to workspace via data(hmp1_2_metaphlan)}
\format{data.frame}
\usage{
hmp1_2_metaphlan
}
\description{
Metaphlan2 HMP1-II
Can be loaded to workspace via data(hmp1_2_metaphlan)
}
\references{
\url{http://github.com/philippmuench}
}
\keyword{datasets}
