% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humann2_tools.R
\name{makeHumann2Barplot}
\alias{makeHumann2Barplot}
\title{makeHumann2Barplot}
\usage{
makeHumann2Barplot(dat, last.plot.colors, scale = "sqrt",
  use.random.colors = T, hide.legend = T, space = "free")
}
\arguments{
\item{dat}{table holding preprocessed humann2 information using `humann2Barplot`}

\item{last.plot.colors}{dataframe of plot colors}

\item{scale}{how to scale the height of bars, on default sqrt}

\item{use.random.colors}{use randomcoloR instead of RColorBrewer}

\item{hide.legend}{boolean information if ledgend should be included}

\item{space}{free or fixed (x scale)}
}
\description{
Generates barplots statified by taxon and metadata
}
