% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansByWat.r
\name{meansByWat}
\alias{meansByWat}
\title{Summarize PAWMAP data by watershed}
\usage{
meansByWat(dfm, vbl, select = NULL, selVar = "metric_code",
  storm = FALSE, sort = "geomean")
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{vbl}{The name of the field storing the value to be summarized}

\item{select}{Select the proper variable from long format}

\item{selVar}{the field containing the metric code}

\item{storm}{Calculate separate means for storm and seasonal data?}

\item{sort}{The statistic to sort the data on}
}
\value{
A table with various means by watershed
}
\description{
Summarize PAWMAP data by watershed
}
