% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/met.cod.r
\name{met.cod}
\alias{met.cod}
\title{Metric codes, names, units, etc for PAWMAP variables.}
\format{A data frame with 688 rows and 7 variables:
\describe{
  \item{metric_code}{short abbreviation for the variable}
  \item{metric_name}{Name of the variable}
  \item{metric_class}{Class of variable: water quality, habitat, biological}
  \item{metric_subclass}{further categorization}
  \item{label}{labeling for graph}
  \item{short}{shorter version of label, where needed}
  \item{units}{units for habitat variables}
}}
\description{
A dataset metric codes, names, units, graph labels and other information for using PAWMAP data.
}

