% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm.plot.r
\name{norm.plot}
\alias{norm.plot}
\title{Plot a variable and test for normality}
\usage{
norm.plot(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
the results of a Shapiro-Wilk normality test
}
\description{
Plot a variable in a qqnorm plot and return the results of a Shapiro-Wilk normality test.
}
\examples{
x <- rnorm(100)
norm.plot(x)
}

