% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.plot.r
\name{panel.cor}
\alias{panel.cor}
\title{A Panel Function for Pairs to Add Correlation Coefficients}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{x, y}{numeric vectors}

\item{digits}{number of significant digits to plot}

\item{prefix}{text to prefix the coefficients}

\item{cex.cor}{character expansion factor}
}
\value{
correlation coefficients for pairs
}
\description{
Panel functions to add correlation coefficients to half a pairs plot
}
\examples{
pairs(USJudgeRatings, lower.panel = panel.smooth, upper.panel = panel.cor)
}
