% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBII_ByWat.r
\name{plotBII_ByWat}
\alias{plotBII_ByWat}
\title{Plot PAWMAP Bird Integrity Index by watershed.}
\usage{
plotBII_ByWat(vbl, dfm = bii)
}
\arguments{
\item{vbl}{the name of the water quality variable to plot}

\item{dfm}{The data frame containing the variable}
}
\value{
A ggplot box plot of the index by watershed
}
\description{
Plot PAWMAP Bird Integrity Index by watershed.
}
\examples{
library(ggplot2)
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_code='bii',
                result=rnorm(length(stationInfo$loc_code)))
d <- mergeStatInfo(d)
p <- plotBII_ByWat('result', d)
p + ggtitle('Bird Integrity Index - Generated Data for Example\\n')
}

