% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBio_InWat.r
\name{plotBio_InWat}
\alias{plotBio_InWat}
\title{Plot PAWMAP biological metrics within a selected watershed.}
\usage{
plotBio_InWat(dfm, vbl, wat, bio = c("bugs", "birds", "fish"), ...)
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{vbl}{the field containing the metric to plot}

\item{wat}{the watershed to plot}

\item{bio}{type of biological data to plot.  Options: 'bugs', 'birds', 'fish'}

\item{...}{Additional arguments to control aesthetic of plotted points (must be string)}
}
\value{
A ggplot dot plot of the Macroinvertebrate Observed/Expected scores
within a watershed
}
\description{
Plot PAWMAP biological metrics within a selected watershed.
}
\examples{
library(ggplot2)
site_identifier <- unique(stationInfo$site_identifier[stationInfo$duration=='P'])
d <- data.frame(site_identifier=unique(site_identifier), metric_name='oep5',
                 watershed='Tryon Creek', result=rnorm(length(site_identifier)))
d <- mergeStatInfo(d)
p <- plotBio_InWat(d, 'result', 'Johnson Creek', 'bugs')
p + ggtitle('Macroinvertebrate Observed/Expected - Generated Data for Example\n')
}
