% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDens.r
\name{plotDens}
\alias{plotDens}
\title{Plot Density of a Variable}
\usage{
plotDens(dat, vari, xtrn = "", print = FALSE)
}
\arguments{
\item{dat}{The data frame containing the variable}

\item{vari}{The variable to plot.  A character string.}

\item{xtrn}{Data transformation}
}
\value{
A density plot
}
\description{
Plot a density and strip plot to illustrate the distribution of a variable
}
\examples{
plotDens(wq14[wq14$metric_code == 'cu', ], 'result', xtrn='log10')
}

