% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFishSppComp.r
\name{plotFishSpp}
\alias{plotFishSpp}
\title{Plot PAWMAP Fish Species Composition Data.}
\usage{
plotFishSpp(dfm, speciesIn = "comm_name", sppLook = "Common_Name")
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{speciesIn}{The field in dfm that identifies species}

\item{sppLook}{Which field to look up the species by: "Common_Name" (default)
or "Species_Name"}
}
\value{
A tornado plot of Fish species presence (number of surveys on which
lwas detected), and abundance (total number of individuals captured)
}
\description{
Plot PAWMAP Fish Species Composition Data.
}

