% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFishSppComp2.r
\name{plotFishSpp2}
\alias{plotFishSpp2}
\title{Plot PAWMAP Fish Species Composition Data.}
\usage{
plotFishSpp2(
  dfm,
  speciesIn = "common_name",
  sppLook = "Common_Name",
  by = "row",
  countFields = c("totNum", "numSurv"),
  limitLength = TRUE,
  ...
)
}
\arguments{
\item{dfm}{The data frame containing fish counts}

\item{speciesIn}{The field in dfm that identifies species}

\item{sppLook}{Look up species traits by: "Common_Name" (default)
or "Species_Name"}

\item{by}{Are the data formatted one row per individual ('row'), or counts per
species ('counts')?}

\item{countFields}{Only used if by='count'. A vector of length 2 indicating
the fields for total number of each species captured first, and the number of
surveys on which they were captured second.}

\item{limitLength}{Should the number of species plotted be limited to the top 25?
(to prevent an overcrowded plot)}

\item{...}{Additional arguments passed to tallyFishSpp}
}
\value{
A tornado plot of fish species presence (number of surveys on which
species was captured), and abundance (total number of individuals captured)
}
\description{
Plot PAWMAP Fish Species Composition Data.
}
