% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGGjitt_ByWat.r
\name{plotGGjitt_ByWat}
\alias{plotGGjitt_ByWat}
\title{Plot A GGplot Jitter plot by Watershed}
\usage{
plotGGjitt_ByWat(dfm, vbl)
}
\arguments{
\item{dfm}{A data frame with a variable to plot}

\item{vbl}{The variable to plot by watershed}
}
\value{
A ggplot graph with the variable on the horizontal axis and the
watersheds along the vertical axis sorted by mean of the variable
}
\description{
Plot A GGplot Jitter plot by Watershed
}
