% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGGjittByWat.r
\name{plotGGjittByWat}
\alias{plotGGjittByWat}
\title{Plot A GGplot Jitter plot by Watershed}
\usage{
plotGGjittByWat(dfm, x, label)
}
\arguments{
\item{dfm}{A data frame with a variable to plot}

\item{x}{The variable to plot by watershed}

\item{label}{A string to label the variable axis}
}
\value{
A ggplot graph with the variable on the horizontal axis and the
watersheds along the vertical axis sorted by mean of the variable
}
\description{
Plot A GGplot Jitter plot by Watershed
}
\examples{
dfm <- data.frame(loc_code=stationInfo$loc_code,
                   vari=rnorm(length(stationInfo$loc_code)))
p <- plotGGjittByWat(dfm, vari, "\\nA generated variable")
p
#  The graph object can be modified
p + ggtitle('An Example Graph\\n')
}

