% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHab_ByWat.r
\name{plotHab_ByWat}
\alias{plotHab_ByWat}
\title{Plot PAWMAP habitat data by watershed.}
\usage{
plotHab_ByWat(dfm, vbl, add.wat = TRUE)
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{vbl}{the field containing the variable to plot}

\item{add.wat}{Should a field designating watershed be added? (Based on station #)}
}
\value{
A ggplot box plot of the variable by watershed
}
\description{
Plot PAWMAP habitat data by watershed.
}
\examples{
library(ggplot2)
## NEED TO UPDATE
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_code='xcl',
                 result=rnorm(length(stationInfo$loc_code)))
p <- plotHab_ByWat(d, 'result')
p + ggtitle('Large Riparian Canopy - Generated Data for Example\n')
}
