% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotOep5_ByWat.R
\name{plotOep5_ByWat}
\alias{plotOep5_ByWat}
\title{Plot PAWMAP Macroinvertebrate Scores by watershed.}
\usage{
plotOep5_ByWat(vbl, dfm = oep5)
}
\arguments{
\item{vbl}{the name of the macroinvertebrate variable to plot}

\item{dfm}{The data frame containing the variable}
}
\value{
A ggplot box plot of the variable by watershed
}
\description{
Plot PAWMAP Macroinvertebrate Scores by watershed.
}
\examples{
library(ggplot2)
stations <- unique(stationInfo$loc_code[stationInfo$duration=='P'])
d <- data.frame(loc_code=unique(stations), metric_code='oe',
                 result=rnorm(length(stations)))
d <- mergeStatInfo(d)
p <- plotOep5_ByWat('oe', d)
p + ggtitle('Macroinvertebrate Observed/Expected Ratio - Generated Data for Example\\n')
}

