% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotOep5_InWat.R
\name{plotOep5_InWat}
\alias{plotOep5_InWat}
\title{Plot PAWMAP Macroinvertebrate Data within a selected watershed.}
\usage{
plotOep5_InWat(wat, dfm = oep5)
}
\arguments{
\item{wat}{the watershed to plot}

\item{dfm}{The data frame containing the variable}
}
\value{
A ggplot dot plot of the Macroinvertebrate Observed/Expected scores
within a watershed
}
\description{
Plot PAWMAP Macroinvertebrate Data within a selected watershed.
}
\examples{
library(ggplot2)
stations <- unique(stationInfo$loc_code[stationInfo$duration=='P'])
d <- data.frame(loc_code=unique(stations), metric_name='oep5',
                 watershed='Tryon Creek', result=rnorm(length(stations)))
d <- mergeStatInfo(d)
p <- plotOep5_InWat('Johnson Creek', d)
p + ggtitle('Macroinvertebrate Observed/Expected - Generated Data for Example\\n')
}

