% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotOep5_wat.R
\name{plotOep5_wat}
\alias{plotOep5_wat}
\title{Plot PAWMAP Macroinvertberate Scores by watershed.}
\usage{
plotOep5_wat(oe, dfm = oep5)
}
\arguments{
\item{dfm}{The data frame conatining the variable}

\item{vbl}{the name of the macroinvertebrate variable to plot}
}
\value{
A ggplot box plot of the variable by watershed
}
\description{
Plot PAWMAP Macroinvertberate Scores by watershed.
}
\examples{
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_name='oe',
result=rnorm(length(stationInfo$loc_code))
d <- mergeStatInfo(d)
p <- plotOep5_wat('oe', d)
p + ggtitle('Macroinvertebrate Observed/Expected Ratio - Generated Data for Example\\n')
}

