% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWQ_ByCycle.r
\name{plotWQ_byCycle}
\alias{plotWQ_byCycle}
\title{Plot PAWMAP water quality data by watershed.}
\usage{
plotWQ_byCycle(
  dfm,
  result = "numeric_result",
  analyte_field = "janus_analyte_name",
  analyte_units = "analyte_units",
  storm = "storm",
  stationInfo = TRUE
)
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{result}{the field containing the numeric value}

\item{analyte_field}{the field containing the name of the water quality analyte}

\item{analyte_units}{the field containing the analyte measurement units}

\item{storm}{the field indicating seasonal or storm samples}

\item{stationInfo}{should station information be added to the data?  If not, a field named 'watershed' is required.}
}
\value{
A ggplot box plot of the variable by watershed
}
\description{
Plot PAWMAP water quality data by watershed.
}
\examples{
library(ggplot2)
### NEEDS TO BE REWRITTEN
d <- data.frame(station=unique(stationInfo$station), metric_name='copper',
                result=rlnorm(length(stationInfo$station)), storm='Seasonal',
                analyte_units = 'ug/L')
d <- rbind(d, data.frame(station=unique(stationInfo$station), metric_name='copper',
                      result=2*rlnorm(length(stationInfo$station)), storm='Seasonal',
                analyte_units = 'ug/L'))
p <- plotWQ_ByWat(d, analyte_field = 'metric_name')
p + ggtitle('Copper - Generated Data for Example\n')
}
