% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWQ_byWat.r
\name{plotWQ_byWat}
\alias{plotWQ_byWat}
\title{Plot PAWMAP water quality data by watershed.}
\usage{
plotWQ_byWat(
  dfm,
  result = "numeric_result",
  analyte_field = "analyte_name",
  analyte_units = "analyte_units",
  storm = "storm_affected",
  stationInfo = TRUE
)
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{result}{the name of the water quality variable to plot}

\item{analyte_field}{the field containing the name of the analyte}

\item{analyte_units}{the field containing the analyte measurement units}

\item{storm}{the field indicating whether the sample was collected
during storm or non-storm conditions}

\item{stationInfo}{should station information be added to the data?  If not, a field named 'watershed' is required.}
}
\value{
A ggplot box plot of the variable by watershed
}
\description{
Plot PAWMAP water quality data by watershed.
}
\examples{
library(ggplot2)
d <- data.frame(site_identifier=unique(stationInfo$site_identifier), janus_analyte_name='copper',
                numeric_result=rlnorm(length(stationInfo$site_identifier)), storm_affected='No',
                analyte_units = 'ug/L')
d <- rbind(d, data.frame(site_identifier=unique(stationInfo$site_identifier), janus_analyte_name='copper',
              numeric_result=2*rlnorm(length(stationInfo$site_identifier)), storm_affected='Yes',
              analyte_units = 'ug/L'))
p <- plotWQ_byWat(d)
p + ggtitle('Copper - Generated Data for Example\n')
}
