% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWQ_InWat.R
\name{plotWQ_InWat}
\alias{plotWQ_InWat}
\title{Plot PAWMAP Water Quality Data within a selected watershed.}
\usage{
plotWQ_InWat(analyte, wat, dfm, analyte_field = "janus_analyte_name",
  result = "numeric_result", analyte_units = "analyte_units")
}
\arguments{
\item{analyte}{the name of the analyte to plot}

\item{wat}{the watershed to plot}

\item{dfm}{The data frame containing the variable}

\item{analyte_field}{the field containing the name of the analyte}

\item{result}{the field storing the numeric value}

\item{analyte_units}{the field containing the measurement units}
}
\value{
A ggplot dot plot of the variable within a watershed
}
\description{
Plot PAWMAP Water Quality Data within a selected watershed.
}
\examples{
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_name='copper',
                result=rlnorm(length(stationInfo$loc_code)), season='S')
d <- rbind(d, data.frame(loc_code=unique(stationInfo$loc_code), metric_name='copper',
                 result=2*rlnorm(length(stationInfo$loc_code)), season='T'))
d <- mergeStatInfo(d)
p <- plotWQ_InWat('copper', 'Johnson Creek', d)
p + ggtitle('Copper - Generated Data for Example\\n')
}
