% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotWQbyWshd.r
\name{plotWQbyWshd}
\alias{plotWQbyWshd}
\title{Plot PAWMAP water quality data by watershed.}
\usage{
plotWQbyWshd(vbl, dfm = wq14, vName = "metric_name")
}
\arguments{
\item{vbl}{the name of the water quality variable to plot}

\item{dfm}{The data frame containing the variable}

\item{vName}{Name of the field storing vbl}
}
\value{
A ggplot box plot of the variable by watershed
}
\description{
Plot PAWMAP water quality data by watershed.
}
\examples{
library(ggplot2)
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_name='copper',
                result=rlnorm(length(stationInfo$loc_code)), season='S')
d <- rbind(d, data.frame(loc_code=unique(stationInfo$loc_code), metric_name='copper',
                      result=2*rlnorm(length(stationInfo$loc_code)), season='T'))
d <- mergeStatInfo(d)
p <- plotWQbyWshd('copper', d)
p + ggtitle('Copper - Generated Data for Example\\n')
}

