% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodge_v.r
\name{pos_fillv}
\alias{pos_fillv}
\title{Stack overlapping intervals and set height to 1.
Assumes that each set has the same horizontal position.}
\usage{
pos_fillv(df, height)
}
\description{
Stack overlapping intervals and set height to 1.
Assumes that each set has the same horizontal position.
}
