% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodge_v.r
\name{pos_stackv}
\alias{pos_stackv}
\title{Stack overlapping intervals.
Assumes that each set has the same horizontal position}
\usage{
pos_stackv(df, height)
}
\description{
Stack overlapping intervals.
Assumes that each set has the same horizontal position
}
