% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFishSurvTrans.r
\name{queryFishEvents}
\alias{queryFishEvents}
\title{Query number of PAWMAP fish surveys and transects at each station.}
\usage{
queryFishEvents()
}
\description{
Query number of PAWMAP fish surveys and transects at each station.
}
