% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFishSurvTrans.r
\name{queryFullFishEvent}
\alias{queryFullFishEvent}
\title{Query number of PAWMAP fish surveys and transects.}
\usage{
queryFullFishEvent()
}
\description{
Query number of PAWMAP fish surveys and transects.
}
