% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFishIbi.R
\name{removeNotFished}
\alias{removeNotFished}
\title{Remove fish surveys that were not attempted}
\usage{
removeNotFished(df, field = "fished")
}
\arguments{
\item{df}{data frame of fish surveys}

\item{field}{name of field indicating whether survey was attempted}
}
\value{
A data frame with non-surveys removed
}
\description{
Remove fish surveys that were not attempted
}
