% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetMixMod.r
\name{runMetMixMod}
\alias{runMetMixMod}
\title{Run a Mixed Model for Metals and TSS in Streams.}
\usage{
runMetMixMod(dat, vrbl)
}
\arguments{
\item{dat}{A data frame with metals and TSS data.}

\item{vrbl}{A string of the name of the field storing variable names}
}
\value{
A list of lme objects for metals
}
\description{
Run a Mixed Model for Metals and TSS in Streams.
}

