% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveMMMranEff.r
\name{saveMMMranEff}
\alias{saveMMMranEff}
\title{Save and Plot the Random Effects from a Mixed Model.}
\usage{
saveMMMranEff(mods, plot = FALSE)
}
\arguments{
\item{mods}{A list of lme objects from a mixed model.}

\item{plot}{Should the random effects be ranked and plotted?}
}
\value{
A data frame of random effects, or a list with data frame and plots
}
\description{
Save and Plot the Random Effects from a Mixed Model.
}
\examples{
mods <- runMetMixMod(wq14, 'metric_name')
saveMMMranEff(mods)  # a data frame
saveMMMranEff(mods, plot = TRUE)  # a list with data frame and plots
}

