% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFishIbi.R
\name{selectFishFields}
\alias{selectFishFields}
\title{Choose the fields to keep in fish survey data}
\usage{
selectFishFields(
  df,
  fields = c("site_identifier", "collection_start_date", "fished",
    "fished_none_collected", "common_name", "fish_length_mm")
)
}
\arguments{
\item{df}{data frame from fish database query}

\item{fields}{a vector of field names to select}
}
\description{
Choose the fields to keep in fish survey data
}
