% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_names.R
\name{shortenTaxons}
\alias{shortenTaxons}
\title{shortenTaxon}
\usage{
shortenTaxons(dat, level = 2, delim = "|")
}
\arguments{
\item{dat}{character vector holding taxonomic names}

\item{level}{column that holds taxonomic level of interest}

\item{delim}{delimiter that seperates levels (e.g. |)
information}
}
\description{
shortens taxon names
}
