% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{std.lns}
\alias{std.lns}
\title{Standard values for plotting}
\format{
A data frame with 7 rows and 3 variables:
\describe{
  \item{metric_code}{code from met.cod table}
  \item{red.line}{Higher (more lenient) standard}
  \item{grn.line}{Lower (more restrictive) standard}

  ...
}
}
\usage{
std.lns
}
\description{
A table of habitat, water quality and biological standrads and benchmarks for plotting.
}
\keyword{datasets}
