% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.r
\name{stLook}
\alias{stLook}
\title{Look up information on a station by number}
\usage{
stLook(station, field = "station")
}
\arguments{
\item{station}{A string or vector of station numbers to look up}

\item{field}{use station (default) or loc_code?}
}
\description{
Look up information on a station by number
}
\examples{
stLook('P2000')
stLook(c('P2000', 'P0012'))
stLook('0012', 'loc_code')
}
