% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substCritStat.r
\name{substCritList}
\alias{substCritList}
\title{List stations exceeding substrate criteria}
\usage{
substCritList(dfm = NULL, wat = NULL)
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{wat}{Optional argument to limit results to a single watershed}
}
\value{
A list of stations exceeding each substrate criterion
}
\description{
List stations exceeding substrate criteria
}
