% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substCritStat.r
\name{substCritStat}
\alias{substCritStat}
\title{Compile total exceedances of substrate criteria by watershed}
\usage{
substCritStat(dfm = NULL)
}
\arguments{
\item{dfm}{The data frame containing the variable}
}
\value{
A table of number of exceedances of each substrate criterion
}
\description{
Compile total exceedances of substrate criteria by watershed
}

