% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFishSurvTrans.r
\name{sumFishSurvTrans}
\alias{sumFishSurvTrans}
\title{Summarize total fish surveys and transects}
\usage{
sumFishSurvTrans(printHist = FALSE)
}
\description{
Summarize total fish surveys and transects
}
