% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallyFishSpp.r
\name{tallyFishSpp}
\alias{tallyFishSpp}
\title{Tally PAWMAP Fish Species Composition Data.}
\usage{
tallyFishSpp(
  dfm,
  speciesIn = "common_name",
  sppLook = "Common_Name",
  by = "row",
  countFields = c("totNum", "numSurv"),
  station = "site_identifier",
  dateField = "collection_start_date",
  noFish = "fished_none_collected"
)
}
\arguments{
\item{dfm}{The data frame containing fish counts}

\item{speciesIn}{The field in dfm that identifies species}

\item{sppLook}{Look up species traits by: "Common_Name" (default)
or "Species_Name"}

\item{by}{Are the data formatted one row per individual ('row'), or counts per
species ('counts')?}

\item{countFields}{Only used if by='count'. A vector of length 2 indicating
the fields for total number of each species captured first, and the number of
surveys on which they were captured second.}

\item{station}{The field identifying the sampling station}

\item{dateField}{The field identifying the sampling date}

\item{noFish}{String indicating the field indicating the survey found no fish}
}
\value{
A table of fish species presence (number of surveys on which
species was captured), and abundance (total number of individuals captured)
}
\description{
Tally PAWMAP Fish Species Composition Data.
}
