% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.r
\name{dropAmphibs}
\alias{dropAmphibs}
\title{Drop Amphibians from a data frame of fish names}
\usage{
dropAmphibs(df, vbl)
}
\arguments{
\item{df}{The data frame containing the fish data}

\item{vbl}{A field within df containing the fish names}
}
\value{
A data frame without amphibians
}
\description{
Drop Amphibians from a data frame of fish names
}
