% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fish.traits}
\alias{fish.traits}
\title{Fish traits table}
\format{
A data frame with 67 rows and 10 variables:
\describe{
  \item{Family}{Taxonomic family}
  \item{Species_Name}{Scientific name}
  \item{Origin}{Native (N) or Alien (A) species?}

  ...
}
}
\usage{
fish.traits
}
\description{
A table of fish traits used in biotic indices.
}
\keyword{datasets}
