% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formSubstrate.R
\name{formSubstrate}
\alias{formSubstrate}
\title{Load and format PAWMAP substrate data}
\usage{
formSubstrate(dfm = hab14, load = FALSE, path = NULL)
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{load}{should the data be loaded from a file?}

\item{path}{path for data if load=TRUE}
}
\value{
A data frame with substrate data in wide format
}
\description{
Load and format PAWMAP substrate data
}

