% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humann2_table}
\alias{humann2_table}
\title{HMANn2 output of Cas gene abundnace amoung HMP1-II samples
Can be loaded to workspace via data("humann2_table")}
\format{data.frame}
\usage{
humann2_table
}
\description{
HMANn2 output of Cas gene abundnace amoung HMP1-II samples
Can be loaded to workspace via data("humann2_table")
}
\references{
\url{http://github.com/philippmuench}
}
\keyword{datasets}
