% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humann2_tools.R
\name{makeHumann2Barplot}
\alias{makeHumann2Barplot}
\title{makeHumann2Barplot}
\usage{
makeHumann2Barplot(dat, scale = "sqrt", palette = "Set1",
  hide.legend = T)
}
\arguments{
\item{dat}{table holding preprocessed humann2 information using `humann2Barplot`}

\item{scale}{how to scale the height of bars, on default sqrt}

\item{palette}{ggplot2 color palette}

\item{hide.legend}{boolean information if ledgend should be included}
}
\description{
Generates barplots statified by taxon and metadata
}
