% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeStatInfo.r
\name{mergeStatInfo}
\alias{mergeStatInfo}
\title{Merge PAWMAP station information into a data frame.}
\usage{
mergeStatInfo(df, fields = c("station", "watershed", "subwat", "loc.lbl",
  "panel", "duration"), ...)
}
\arguments{
\item{df}{A data frame with a station field}

\item{fields}{Fields to add from the station table}

\item{by.x}{The station field within the stationInfo table}

\item{by.y}{The station field within the data frame df}
}
\value{
The original data frame merged with station fields.  The merge
returns all rows of df even if there is no match, to indicate that there
are stations in df not present in the station table.
}
\description{
Merge PAWMAP station information into a data frame.
}
\examples{
df <- data.frame(loc_code=unique(stationInfo$station),
                 result=rnorm(length(stationInfo$station)))
mergeStatInfo(df)
}
