% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeStatInfo.r
\name{mergeStatInfo}
\alias{mergeStatInfo}
\title{Merge PAWMAP station information into a data frame.}
\usage{
mergeStatInfo(
  df,
  by.y = "site_identifier",
  filterNA = TRUE,
  fields = c("site_identifier", "watershed", "subwat", "loc.lbl", "panel", "duration"),
  renameStat = NULL
)
}
\arguments{
\item{df}{A data frame with a station field}

\item{by.y}{Name of station field, to be matched with site_identifier}

\item{filterNA}{should stations w/o info be filtered? (A warning listing the
unmatched stations is provided.)}

\item{fields}{Fields to add from the station table}

\item{renameStat}{Should the site_identifier field be renamed to station?
(to match historical output)}
}
\value{
The original data frame merged with station info fields.  The merge
returns all rows of df even if there is no match, to indicate that there
are stations in df not present in the station table.
}
\description{
Merge PAWMAP station information into a data frame.
}
\examples{
df <- data.frame(site_identifier=unique(stationInfo$site_identifier),
                 result=rnorm(length(stationInfo$site_identifier)))
mergeStatInfo(df)
}
