% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeStatInfo.r
\name{mergeStatInfo}
\alias{mergeStatInfo}
\title{Merge PAWMAP station information into a data frame.}
\usage{
mergeStatInfo(dat, stat = "loc_code", fields = c("watershed", "subwat",
  "loc.lbl", "panel", "duration"))
}
\arguments{
\item{dat}{A data frame with station field.}

\item{stat}{The station field within dat}

\item{fields}{Fields to add from the station table}
}
\value{
The original data frame merged with station fields
}
\description{
Merge PAWMAP station information into a data frame.
}
\examples{
mergeStatInfo(wq14[wq14$metric_code == 'cu', ])
}

