% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeStatInfo.r
\name{mergeStatInfo}
\alias{mergeStatInfo}
\title{Merge PAWMAP station information into a data frame.}
\usage{
mergeStatInfo(
  df,
  fields = c("station", "watershed", "subwat", "loc.lbl", "panel", "duration"),
  ...
)
}
\arguments{
\item{df}{A data frame with a station field}

\item{fields}{Fields to add from the station table}

\item{...}{Provides the ability to add by (if using loc_code) or by.x and
by.y if the station field is not named "station"}
}
\value{
The original data frame merged with station info fields.  The merge
returns all rows of df even if there is no match, to indicate that there
are stations in df not present in the station table.
}
\description{
Merge PAWMAP station information into a data frame.
}
\examples{
df <- data.frame(station=unique(stationInfo$station),
                 result=rnorm(length(stationInfo$station)))
mergeStatInfo(df)
}
