% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBio_ByWat.r
\name{plotBio_ByWat}
\alias{plotBio_ByWat}
\title{Plot A GGplot jitter plot of a biological metric by watershed}
\usage{
plotBio_ByWat(dfm, vbl)
}
\arguments{
\item{dfm}{A data frame with a variable to plot}

\item{vbl}{The variable to plot by watershed}
}
\value{
A ggplot graph with the variable on the horizontal axis and the
watersheds along the vertical axis sorted by mean of the variable
}
\description{
Plot A GGplot jitter plot of a biological metric by watershed
}

