% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBio_InWat.r
\name{plotBio_InWat}
\alias{plotBio_InWat}
\title{Plot PAWMAP biological metrics within a selected watershed.}
\usage{
plotBio_InWat(dfm, vbl, wat, bio)
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{vbl}{the field containing the metric to plot}

\item{wat}{the watershed to plot}
}
\value{
A ggplot dot plot of the Macroinvertebrate Observed/Expected scores
within a watershed
}
\description{
Plot PAWMAP biological metrics within a selected watershed.
}
\examples{
library(ggplot2)
stations <- unique(stationInfo$loc_code[stationInfo$duration=='P'])
d <- data.frame(loc_code=unique(stations), metric_name='oep5',
                 watershed='Tryon Creek', result=rnorm(length(stations)))
d <- mergeStatInfo(d)
p <- plotBio_InWat(d, 'result', 'Johnson Creek')
p + ggtitle('Macroinvertebrate Observed/Expected - Generated Data for Example\\n')
}
