% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCDF_byWat.r
\name{plotCDF_byWat}
\alias{plotCDF_byWat}
\title{Plot Empirical Cumulative Distribution Plots of PAWMAP data by watershed.}
\usage{
plotCDF_byWat(vbl, dfm, season = "seasonal", scale = "linear")
}
\arguments{
\item{vbl}{the name of the water quality variable to plot}

\item{dfm}{The data frame containing the variable}

\item{season}{Should the seasonal or storm data be plotted?}
}
\value{
A ggplot ECDF plot of the variable by watershed
}
\description{
Plot Empirical Cumulative Distribution Plots of PAWMAP data by watershed.
}

