% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDens.r
\name{plotDens}
\alias{plotDens}
\title{Plot Density Distribution of a Variable}
\usage{
plotDens(df, vbl, xtran = NULL, title = NULL, print = FALSE)
}
\arguments{
\item{df}{The data frame containing the variable}

\item{vbl}{The variable to plot.  A character string.}

\item{xtran}{ggplot data transformation}

\item{title}{Title for the plot}

\item{print}{Should the graph be saved to a file?}
}
\value{
A density plot
}
\description{
Plot a density and strip plot to illustrate the distribution of a variable
}
\examples{
df <- data.frame(x=rlnorm(100))
plotDens(df, 'x')
plotDens(df, 'x', xtran = 'log10')
}
