% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHab_ByWat.r
\name{plotHab_ByWat}
\alias{plotHab_ByWat}
\title{Plot PAWMAP habitat data by watershed.}
\usage{
plotHab_ByWat(vbl, dfm = hab14)
}
\arguments{
\item{vbl}{the name of the water quality variable to plot}

\item{dfm}{The data frame containing the variable}
}
\value{
A ggplot box plot of the variable by watershed
}
\description{
Plot PAWMAP habitat data by watershed.
}
\examples{
library(ggplot2)
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_code='xcl',
                 result=rnorm(length(stationInfo$loc_code)))
p <- plotHab_ByWat('xcl', d)
p + ggtitle('Large Riparian Canopy - Generated Data for Example\\n')
}
