% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHab_InWat.R
\name{plotHab_InWat}
\alias{plotHab_InWat}
\title{Plot PAWMAP Habitat Data within a selected watershed.}
\usage{
plotHab_InWat(dfm, vbl, wat, metric_code = NULL)
}
\arguments{
\item{dfm}{The data frame containing the habitat data}

\item{vbl}{The field containing the numeric habitat values to plot}

\item{wat}{The watershed to plot}

\item{metric_code}{The function assumes the field name is the metric code.
If not, (e.g., habitat data are in a field named "result") supply the
metric code to allow conversions and benchmarks}
}
\value{
A ggplot dot plot of the variable within a watershed
}
\description{
Plot PAWMAP Habitat Data within a selected watershed.
}
\examples{
library(ggplot2)
d <- data.frame(site_identifier=unique(stationInfo$site_identifier), metric_code='xcl',
                 result=rnorm(length(stationInfo$site_identifier)))
d <- mergeStatInfo(d)
p <- plotHab_InWat(d, 'result', 'Willamette Streams')
p + ggtitle('Riparian Canopy - Generated Data for Example\n')
}
