% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHab_InWat.r
\name{plotHab_InWat}
\alias{plotHab_InWat}
\title{Plot PAWMAP Habitat Data within a selected watershed.}
\usage{
plotHab_InWat(vbl, wat, dfm, metric_code = NULL)
}
\arguments{
\item{vbl}{the field containing the variable to plot}

\item{wat}{the watershed to plot}

\item{dfm}{The data frame containing the variable}

\item{metric_code}{The function assumes the field name is the metric code.  If not,
supply the metric code to allow conversions and benchmarks}
}
\value{
A ggplot dot plot of the variable within a watershed
}
\description{
Plot PAWMAP Habitat Data within a selected watershed.
}
\examples{
library(ggplot2)
## NEEDS TO BE UPDATED
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_code='xcl',
                 watershed='Willamette Streams', result=rnorm(length(stationInfo$loc_code)))
d <- mergeStatInfo(d)
p <- plotHab_InWat('xcl', 'Willamette Streams', d)
p + ggtitle('Riparian Canopy - Generated Data for Example\\n')
}
