% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotHab_inwat.R
\name{plotHab_inwat}
\alias{plotHab_inwat}
\title{Plot PAWMAP Habitat Data within a selected watershed.}
\usage{
plotHab_inwat(vbl, wat, dfm = hab14)
}
\arguments{
\item{vbl}{the name of the variable to plot}

\item{wat}{the watershed to plot}

\item{dfm}{The data frame conatining the variable}
}
\value{
A ggplot dot plot of the variable within a watershed
}
\description{
Plot PAWMAP Habitat Data within a selected watershed.
}
\examples{
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_name='xcl',
result=rnorm(length(stationInfo$loc_code))
d <- mergeStatInfo(d)
p <- plotHab_inwat('xcl', d)
p + ggtitle('Riparian Canopy - Generated Data for Example\\n')
}

