% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHab_InWat.r
\name{plotHab_InWat}
\alias{plotHab_InWat}
\title{Plot PAWMAP Habitat Data within a selected watershed.}
\usage{
plotHab_InWat(vbl, wat, dfm = hab14)
}
\arguments{
\item{vbl}{the name of the variable to plot}

\item{wat}{the watershed to plot}

\item{dfm}{The data frame conatining the variable}
}
\value{
A ggplot dot plot of the variable within a watershed
}
\description{
Plot PAWMAP Habitat Data within a selected watershed.
}
\examples{
library(ggplot2)
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_code='xcl',
                 watershed='Willamette Streams', result=rnorm(length(stationInfo$loc_code)))
d <- mergeStatInfo(d)
p <- plotHab_InWat('xcl', 'Willamette Streams', d)
p + ggtitle('Riparian Canopy - Generated Data for Example\\n')
}
