% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIbi_InWat.R
\name{plotIbi_InWat}
\alias{plotIbi_InWat}
\title{Plot PAWMAP Fish Data within a selected watershed.}
\usage{
plotIbi_InWat(wat, dfm = runFishIbi(), onlySummer = FALSE)
}
\arguments{
\item{wat}{the watershed to plot}

\item{dfm}{The data frame containing the variable}
}
\value{
A ggplot dot plot of Fish Index of Biotic Integrity scores
within a watershed
}
\description{
Plot PAWMAP Fish Data within a selected watershed.
}
\examples{
library(ggplot2)
stations <- unique(stationInfo$loc_code[stationInfo$duration=='P'])
d <- data.frame(loc_code=unique(stations), metric_name='oep5',
                 watershed='Johnson Creek', result=rnorm(length(stations)))
d <- mergeStatInfo(d)
p <- plotIbi_InWat(d, 'Johnson Creek')
p + ggtitle('Fish Index of Biotic Integrity - Generated Data for Example\n')
}
