% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotOep5_inwat.R
\name{plotOep5_inwat}
\alias{plotOep5_inwat}
\title{Plot PAWMAP Macroinvertebrate Data within a selected watershed.}
\usage{
plotOep5_inwat(wat, dfm = oep5)
}
\arguments{
\item{wat}{the watershed to plot}
}
\value{
A ggplot dot plot of the Macroinvertebrate Observed/Expected scores
within a watershed
}
\description{
Plot PAWMAP Macroinvertebrate Data within a selected watershed.
}
\examples{
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_name='oep5',
watershed='Tryon Creek', result=rnorm(length(stationInfo$loc_code))
d <- mergeStatInfo(d)
p <- plotOep5_inwat(d)
p + ggtitle('Macroinvertebrate Observed/Expected - Generated Data for Example\\n')
}

