% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRespVsPred.r
\name{plotRespVsPred}
\alias{plotRespVsPred}
\title{Plot Correlations of Predictors vs. a Response}
\usage{
plotRespVsPred(dfm, resp, pred)
}
\arguments{
\item{dfm}{the dataset containing the predictor and response variables}

\item{resp}{the name of the response variable within the data frame, as a string}

\item{pred}{a vector of the column numbers of the predictors or the names as strings}
}
\value{
Faceted scatterplots of each predictor vs. the response
}
\description{
Takes a response variable (e.g., a macroinvertebrate or fish IBI) and
plots it against a number of predictor variables.
}
\details{
Creates faceted scatterplots with each predictor as a facet, the
correlation coefficient plotted in the corner, and the facets sorted in
decreasing correlation strength.
}
\examples{
plotRespVsPred(mtcars, 'mpg', c('hp', 'wt', 'disp'))
}
