% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSubstrate.r
\name{plotSubstrate}
\alias{plotSubstrate}
\title{Plot PAWMAP substrate data}
\usage{
plotSubstrate(substrate, dfm = NULL, load = TRUE, path = NULL)
}
\arguments{
\item{substrate}{plot "fines" or "gravel"?}

\item{dfm}{The data frame containing the variable}

\item{load}{should the data be loaded from a file?}

\item{path}{path for data if load=TRUE}
}
\value{
A ggplot dot plot of the substrate data plotted vs. percent riffles,
with points above relevant criteria highlighted and labeled
}
\description{
Plot PAWMAP substrate data
}
\examples{
num <- length(stationInfo$site_identifier)
dfm <- data.frame(site_identifier=stationInfo$site_identifier,
                   pct_ri=runif(num, 0, 100),
                   pct_fn=runif(num, 0, 60),
                   pct_grav=runif(num, 0, 60))
dfm <- mergeStatInfo(dfm)
plotSubstrate('fines', dfm)
plotSubstrate('fines', dfm)
}
