% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWQ_byCycle.r
\name{plotWQ_byCycle}
\alias{plotWQ_byCycle}
\title{Plot PAWMAP water quality data by watershed.}
\usage{
plotWQ_byCycle(
  dfm,
  result = "numeric_result",
  analyte_field = "janus_analyte_name",
  analyte_units = "analyte_units",
  storm = "storm",
  stationInfo = TRUE
)
}
\arguments{
\item{dfm}{The data frame containing the variable}

\item{result}{the field containing the numeric value}

\item{analyte_field}{the field containing the name of the water quality analyte}

\item{analyte_units}{the field containing the analyte measurement units}

\item{storm}{the field indicating seasonal or storm samples}

\item{stationInfo}{should station information be added to the data?  If not, a field named 'watershed' is required.}
}
\value{
A ggplot box plot of the variable plotted by sampling cycle
}
\description{
Plot PAWMAP water quality data by watershed.
}
\examples{
library(ggplot2)
stations <- unique(stationInfo$site_identifier)
num_stations <- length(stations)
d <- data.frame(site_identifier=stations, janus_analyte_name='copper',
                numeric_result=rlnorm(num_stations), storm='Seasonal',
                analyte_units = 'ug/L',
                cycle=replicate(num_stations, sample(c(1,2), 1)))
d <- rbind(d, data.frame(site_identifier=stations, janus_analyte_name='copper',
              numeric_result=2*rlnorm(num_stations), storm='Storm',
              analyte_units = 'ug/L',
              cycle=replicate(num_stations, sample(c(1,2), 1))))

p <- plotWQ_byCycle(d)
p + ggtitle('Copper - Generated Data for Example\n')
}
