% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotWQ_inwat.R
\name{plotWQ_inwat}
\alias{plotWQ_inwat}
\title{Plot PAWMAP Water Quality Data within a selected watershed.}
\usage{
plotWQ_inwat(vbl, wat, dfm = wq14, vName = "metric_name")
}
\arguments{
\item{vbl}{the name of the variable to plot}

\item{wat}{the watershed to plot}

\item{dfm}{The data frame conatining the variable}
}
\value{
A ggplot dot plot of the variable within a watershed
}
\description{
Plot PAWMAP Water Quality Data within a selected watershed.
}
\examples{
library(ggplot2)
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_name='copper',
                 result=rlnorm(length(stationInfo$loc_code)))
d <- mergeStatInfo(d)
p <- plotWQ_inwat('copper', 'Johnson Creek', d)
p + ggtitle('Copper - Generated Data for Example\\n')
}

