% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotWQ_InWat.R
\name{plotWQ_InWat}
\alias{plotWQ_InWat}
\title{Plot PAWMAP Water Quality Data within a selected watershed.}
\usage{
plotWQ_InWat(vbl, wat, dfm = wq14, vName = "metric_name")
}
\arguments{
\item{vbl}{the name of the variable to plot}

\item{wat}{the watershed to plot}

\item{dfm}{The data frame containing the variable}

\item{vName}{Name of the field storing vbl}
}
\value{
A ggplot dot plot of the variable within a watershed
}
\description{
Plot PAWMAP Water Quality Data within a selected watershed.
}
\examples{
library(ggplot2)
d <- data.frame(loc_code=unique(stationInfo$loc_code), metric_name='copper',
                result=rlnorm(length(stationInfo$loc_code)), season='S')
d <- rbind(d, data.frame(loc_code=unique(stationInfo$loc_code), metric_name='copper',
                 result=2*rlnorm(length(stationInfo$loc_code)), season='T'))
d <- mergeStatInfo(d)
p <- plotWQ_InWat('copper', 'Johnson Creek', d)
p + ggtitle('Copper - Generated Data for Example\\n')
}

