% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runFishIbi.R
\name{runFishIbi}
\alias{runFishIbi}
\title{Get Fish Index of Biotic Integrity scores}
\usage{
runFishIbi()
}
\value{
A data frame with fish ibi scores and subscores
}
\description{
Get Fish Index of Biotic Integrity scores
}

