% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.r
\name{stLook}
\alias{stLook}
\title{Look up information on a station by number}
\usage{
stLook(station, field = "loc_code")
}
\arguments{
\item{station}{the station number to look up (as a string)}

\item{field}{use loc_code (default) or station?}
}
\description{
Look up information on a station by number
}
\examples{
stLook('2000')
stLook(c('2000', '0012'))
stLook('P0012', 'station')
}

