% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFishSurvTrans.r
\name{sumFishSurvTrans}
\alias{sumFishSurvTrans}
\title{Query number of PAWMAP fish surveys and transects.}
\usage{
sumFishSurvTrans()
}
\description{
Query number of PAWMAP fish surveys and transects.
}
